requirements.txt
pydantic    
llama-index-core
llama-index-embeddings-openai 
llama-index-agent-llm-compiler
llama-index-llms-openai
httpx
termcolor
config
docker
rich
datasets